requirements.txt
torch>=2.0.0
transformers>=4.30.0
unbabel-comet>=2.0.0
huggingface-hub>=0.16.0
